Refactored the profile generation method, now iterating through batches and splitting the method into two based on the type of motion trigger to reduce the number of checks.

Also added a yield after each batch so that we don't block other threads for the entire calculation, causing other parts to fail during configure from CA timeouts.

A couple of rough benchmarks on p45-ws001 for start of row triggering:

| Num points  | Configure time before | Configure time after |
| ---------------- | ----------------------------- | -------------------------- |
| 1,000,000 | 1.5s | 1.1s |
|  15,000,000 | 37s  | 22s |

This fixes #358